Clean live chart handling
