+++ Tunny/WPF/Common/ChartType.cs
@@ -1,0 +1,8 @@
+namespace Tunny.WPF.Common
+{
+    public enum ChartType
+    {
+        Line,
+        Scatter
+    }
+}
