+++ Tunny/WPF/Views/Pages/Optimize/LiveChartPage.xaml
@@ -1,0 +1,67 @@
+﻿<Page
+    x:Class="Tunny.WPF.Views.Pages.Optimize.LiveChartPage"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:local="clr-namespace:Tunny.WPF.Views.Pages.Optimize"
+    xmlns:lvc="clr-namespace:LiveChartsCore.SkiaSharpView.WPF;assembly=LiveChartsCore.SkiaSharpView.WPF"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    xmlns:viewmodels="clr-namespace:Tunny.WPF.ViewModels"
+    Title="LiveChartPage"
+    d:DesignHeight="450"
+    d:DesignWidth="800"
+    mc:Ignorable="d">
+    <Page.DataContext>
+        <viewmodels:LiveChartViewModel />
+    </Page.DataContext>
+
+    <Grid>
+        <Grid.RowDefinitions>
+            <RowDefinition Height="AUTO" />
+            <RowDefinition Height="*" />
+        </Grid.RowDefinitions>
+        <Grid>
+            <Grid.ColumnDefinitions>
+                <ColumnDefinition Width="*" />
+                <ColumnDefinition Width="AUTO" />
+                <ColumnDefinition Width="2*" />
+                <ColumnDefinition Width="AUTO" />
+                <ColumnDefinition Width="2*" />
+            </Grid.ColumnDefinitions>
+
+            <CheckBox
+                HorizontalAlignment="Center"
+                VerticalAlignment="Center"
+                Content="Live Chart"
+                IsChecked="True" />
+            <Label
+                Grid.Column="1"
+                HorizontalAlignment="Center"
+                VerticalAlignment="Center"
+                Content="X Target" />
+            <ComboBox
+                x:Name="ChartXTargetComboBox"
+                Grid.Column="2"
+                Margin="5" />
+            <Label
+                Grid.Column="3"
+                HorizontalAlignment="Center"
+                VerticalAlignment="Center"
+                Content="Y Target" />
+            <ComboBox
+                x:Name="ChartYTargetComboBox"
+                Grid.Column="4"
+                Margin="5" />
+        </Grid>
+
+        <lvc:CartesianChart
+            x:Name="OptimizaLiveChart"
+            Grid.Row="1"
+            Margin="5"
+            AnimationsSpeed="0"
+            Series="{Binding ChartSeries}"
+            XAxes="{Binding ChartXAxes}"
+            YAxes="{Binding ChartYAxes}" />
+
+    </Grid>
+</Page>
